refactor(qwen35): split scheduler backend into scheduler/backend.rs
Move the SingleGpuBackend / TpSchedulerBackend / SchedulerBackend /
AsyncPrefillOutput cluster (553 lines) out of the 2624-line scheduler.rs
into scheduler/backend.rs. Pure move: mod backend + use self::backend::*.

Validated on A100: cargo check + cargo fmt OK, cargo test --lib 95 passed /
0 failed (7 TP tests ignored: need 2 CUDA devices + Qwen3.5 weights).
scheduler.rs 2624 -> 2074 lines.
